Auto-login the public user if the requested non-public page has a public equivalent
All of the non-public to public url space mapping now happens even if
the request is not yet authed as the public user.  By checking if there
is a public equivalent of the request and logging the client in if so,
we allow our autohandler/Default callback do the redirecting later in
the request handling.

This resolves [rt.cpan.org #76301].