Remove the welcome message since the dead-end link is confusing
Remove the 'welcome guest' menu option. It is intended as a reminder to
the current user that they are logged in as a guest. It is confusing,
however, because it renders as a link and the link has no target.

Leave the commented code in place until RT's menus support a no-link option
or a sensible target page is created.